split associativity test into seperate file to reduce test execution time
